+++ smoo-front/src/Routes/Channel.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function Channel() {
+  return <div>channel</div>;
+}
+
+export default Channel;
