+++ smoo-front/src/Routes/Main.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function Main() {
+  return <div>main</div>;
+}
+
+export default Main;
