+++ smoo-front/src/Routes/Workspace.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function Workspace() {
+  return <div>workspace</div>;
+}
+
+export default Workspace;
