+++ smoo-front/src/Routes/index.js
@@ -1,0 +1,5 @@
+// eslint-disable-next-line import/prefer-default-export
+export { default as Login } from './Login';
+export { default as Workspace } from './Workspace';
+export { default as Channel } from './Channel';
+export { default as Main } from './Main';
